% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csfee-package.r, R/data.R
\docType{package}
\name{csfee}
\alias{csfee}
\alias{csfee-package}
\title{csfee.}
\format{A data frame with seven variables.}
\source{
[scholar.princeton.edu/dalexand/data-1]()
}
\usage{
csfee
}
\description{
csfee.

State-year-month level Medicaid Fee Schedules for Birth Procedures
}
\keyword{datasets}

